Created Vehicle Speed detection using openCV and python
This project is to predict the speed of a vehicle with respect to the data from a recorded video source. Traffic congestion is the main problem faced by big cities all over the world. One approach to reduce congestion levels is to improve traffic management that regulates and controls the number of vehicles. To evaluate the impact of traffic management before direct implementation on the highway, traffic modeling can be carried out. Parameters in modeling traffic must be determined from a calibration process where the vehicle is accurately measured for its position and speed. This study aims to propose an efficient calibration procedure with accurate results, based on recorded vehicle movement in perspective view. The proposed system serves to display the process of calibrating camera for traffic simulation to obtain information on the average speed of a vehicle. It is based on Python programming and its libraries, such as OpenCV. The output video is result analysis, which shows the information about the position track of vehicle and its average speed. In the experiment results, the accuracy of vehicle position detection is evaluated.